229 revert table packer changes (#230)
* Revert "Comment all couts #194"

This reverts commit 15067809d6a0e803b505f6d75ed7831e4d06e730.

* Minor function's signatures changes #229

* Delete one of lookup_table_definition.hpp files #229
